requirements.txt
fastapi
uvicorn
tensorflow-cpu
opencv-python-headless
numpy
requests
python-multipart
pillow